+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 - Add Dutch translation #91 (@robinaartsma)
+- Add dynamic layout for the presenter view
 
 ### Fixes and improvements
 
@@ -12,6 +13,7 @@
 - Fix unknown locales
 - Add validation to avoid user to self assign as a facilitator
 - Toggle for message reactions is replaced with toggle for message and global reactions
+- Improve embed integration in presenter view
 
 ## v2.0.0